プロフィールとファイルの関連モデル名変更
