+++ src/Model/Table/ProfileFilesTable.php
@@ -1,0 +1,87 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Model\Table;
+
+use Cake\ORM\RulesChecker;
+use Cake\ORM\Table;
+use Cake\Validation\Validator;
+
+/**
+ * ProfileFiles Model
+ *
+ * @property \App\Model\Table\ProfilesTable&\Cake\ORM\Association\BelongsTo $Profiles
+ * @property \App\Model\Table\FilesTable&\Cake\ORM\Association\BelongsTo $Files
+ * @method \App\Model\Entity\ProfileFile newEmptyEntity()
+ * @method \App\Model\Entity\ProfileFile newEntity(array $data, array $options = [])
+ * @method \App\Model\Entity\ProfileFile[] newEntities(array $data, array $options = [])
+ * @method \App\Model\Entity\ProfileFile get($primaryKey, $options = [])
+ * @method \App\Model\Entity\ProfileFile findOrCreate($search, ?callable $callback = null, $options = [])
+ * @method \App\Model\Entity\ProfileFile patchEntity(\Cake\Datasource\EntityInterface $entity, array $data, array $options = [])
+ * @method \App\Model\Entity\ProfileFile[] patchEntities(iterable $entities, array $data, array $options = [])
+ * @method \App\Model\Entity\ProfileFile|false save(\Cake\Datasource\EntityInterface $entity, $options = [])
+ * @method \App\Model\Entity\ProfileFile saveOrFail(\Cake\Datasource\EntityInterface $entity, $options = [])
+ * @method \App\Model\Entity\ProfileFile[]|\Cake\Datasource\ResultSetInterface|false saveMany(iterable $entities, $options = [])
+ * @method \App\Model\Entity\ProfileFile[]|\Cake\Datasource\ResultSetInterface saveManyOrFail(iterable $entities, $options = [])
+ * @method \App\Model\Entity\ProfileFile[]|\Cake\Datasource\ResultSetInterface|false deleteMany(iterable $entities, $options = [])
+ * @method \App\Model\Entity\ProfileFile[]|\Cake\Datasource\ResultSetInterface deleteManyOrFail(iterable $entities, $options = [])
+ * @mixin \Cake\ORM\Behavior\TimestampBehavior
+ */
+class ProfileFilesTable extends Table
+{
+    /**
+     * Initialize method
+     *
+     * @param array $config The configuration for the Table.
+     * @return void
+     */
+    public function initialize(array $config): void
+    {
+        parent::initialize($config);
+
+        $this->setTable('profile_files');
+        $this->setDisplayField('id');
+        $this->setPrimaryKey('id');
+
+        $this->addBehavior('Timestamp');
+
+        $this->belongsTo('Profiles', [
+            'foreignKey' => 'profile_id',
+            'joinType' => 'INNER',
+        ]);
+        $this->belongsTo('Files', [
+            'foreignKey' => 'file_id',
+            'joinType' => 'INNER',
+        ]);
+    }
+
+    /**
+     * Default validation rules.
+     *
+     * @param \Cake\Validation\Validator $validator Validator instance.
+     * @return \Cake\Validation\Validator
+     */
+    public function validationDefault(Validator $validator): Validator
+    {
+        $validator
+            ->integer('id')
+            ->allowEmptyString('id', null, 'create');
+
+        return $validator;
+    }
+
+    /**
+     * Returns a rules checker object that will be used for validating
+     * application integrity.
+     *
+     * @param \Cake\ORM\RulesChecker $rules The rules object to be modified.
+     * @return \Cake\ORM\RulesChecker
+     */
+    public function buildRules(RulesChecker $rules): RulesChecker
+    {
+        $rules->add($rules->existsIn('profile_id', 'Profiles'), ['errorField' => 'profile_id']);
+        $rules->add($rules->existsIn('file_id', 'Files'), ['errorField' => 'file_id']);
+
+        return $rules;
+    }
+}
